requirements.txt
streamlit==1.56.0
requests==2.33.1
python-dotenv==1.2.2
pandas==3.0.2
urllib3==2.6.3
